+++ gist/api.py
@@ -31,9 +31,8 @@ class GistApi(object):
             return json.loads(open(cache_dir).read())
 
         # If no cache, just request server
-        url = GIST_BASE_URL % "gists"
-        _gists = requests.get(url, headers=self.headers).json()
-        return _gists
+        self.res = requests.get(GIST_BASE_URL % "gists", headers=self.headers)
+        return self.res
 
     def get(self, url):
         self.res = requests.get(url, headers=self.headers)