+++ gist/lib/callback.py
@@ -1,4 +1,5 @@
 import os
+import json
 import sublime
 
 from . import util
@@ -13,9 +14,7 @@ def refresh_gist(res, options):
     with open(file_full_name, "wb") as fp:
         fp.write(res.content)
 
-    Printer.get("log").write("%s update succeed" % filename)
-    sublime.set_timeout_async(Printer.get("log").hide_panel, 
-        settings["delay_seconds_for_hiding_panel"] * 1000)
+    show_message("%s update succeed" % filename)
 
 def open_gist(res, options):
     filename = options["filename"]
@@ -43,14 +42,16 @@ def delete_gist(res, options):
 
     settings = util.get_settings()
 
-    view = util.get_view_by_file_name(file_full_name)
-    if view:
-        sublime.active_window().focus_view(view)
-        sublime.active_window().run_command("close")
+    # Close corresponding view by file full name
+    util.close_view_by_filename(file_full_name)
+
+    # Remove file name
     os.remove(file_full_name)
-    Printer.get("log").write("%s delete succeed" % filename)
-    sublime.set_timeout_async(Printer.get("log").hide_panel, 
-        settings["delay_seconds_for_hiding_panel"] * 1000)
+
+    show_message("%s delete succeed" % filename)
+
+    # Reload gist cache
+    sublime.active_window().run_command('reload_gist_cache')
 
 def create_gist(res, options):
     # Get filename and content
@@ -72,17 +73,46 @@ def create_gist(res, options):
     sublime.active_window().open_file(file_full_name)
 
     # Success message
-    Printer.get("log").write("%s is created successfully" % filename)
-    sublime.set_timeout_async(Printer.get("log").hide_panel, 
-        settings["delay_seconds_for_hiding_panel"] * 1000)
+    show_message("%s is created successfully" % filename)
+
+    # Reload gist cache
+    sublime.active_window().run_command('reload_gist_cache')
 
 def update_gist(res, options):
     # Get file_full_name
     file_full_name = options["file_full_name"]
     base, filename = os.path.split(file_full_name)
+    show_message("%s is update successfully" % filename)
 
+def rename_gist(res, options):
+    # Get file_full_name
+    old_file_full_name = options["file_full_name"]
+    old_filename = options["old_filename"]
+    new_filename = options["new_filename"]
+
+    # Get settings
     settings = util.get_settings()
 
-    Printer.get("log").write("%s is update successfully" % filename)
+    # Close corresponding view by file full name
+    util.close_view_by_filename(old_file_full_name)
+
+    # Rename 
+    new_file_full_name = os.path.join(settings["workspace"], new_filename)
+    os.rename(old_file_full_name, new_file_full_name)
+    sublime.active_window().open_file(new_file_full_name)
+
+    show_message("%s is renamed to %s successfully" % (
+        old_filename, new_filename
+    ))
+
+    # Reload gist cache
+    sublime.active_window().run_command('reload_gist_cache')
+
+def add_gists_to_cache(res, options):
+    util.add_gists_to_cache(res.json())
+
+def show_message(msg):
+    settings = util.get_settings()
+    Printer.get("log").write(msg)
     sublime.set_timeout_async(Printer.get("log").hide_panel, 
         settings["delay_seconds_for_hiding_panel"] * 1000)